libservo: Move network cache management to `NetworkManager`
This patch moves network cache management out of `SiteDataManager` and into
`NetworkManager`, aligning the API with how browser engines conceptually
separate network layer state from site scoped data.

`clear_cache` is now exposed on `NetworkManager`, and documentation has been
added to clarify the responsibility boundaries between site data and network
state.
